Add FT for 'My lists' page and 'add items' helper
The `add_list_items` helper will add an item to a list and assert it
was added in the correct format.

The my lists test has a 'wip' django test tag to make it easer to
include/exclude from the tests and also make it clear that the
functionality it represents is a work in progress.